Sonar commit 2
Second commit for resolving Sonar Cloud issues
